+++ src/pubsub/src/publisher/builder.rs
@@ -188,9 +188,10 @@ impl PublisherBuilder {
     /// ```
     /// # use google_cloud_pubsub::client::Publisher;
     /// # async fn sample() -> anyhow::Result<()> {
-    /// use google_cloud_gax::retry_policy::{AlwaysRetry, RetryPolicyExt};
+    /// use google_cloud_gax::retry_policy::RetryPolicyExt;
+    /// use google_cloud_pubsub::retry_policy::RetryableErrors;
     /// let client = Publisher::builder("projects/my-project/topics/my-topic")
-    ///     .with_retry_policy(AlwaysRetry.with_attempt_limit(3))
+    ///     .with_retry_policy(RetryableErrors.with_attempt_limit(3))
     ///     .build().await?;
     /// # Ok(()) };
     /// ```
